# apps/main/src/spec/openapi.yaml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Most Secure Mailbox
  license:
    name: MIT
servers:
  - url: http://localhost:4000
paths:
  /users:
    post:
      summary: Create a new user
      operationId: createUser
      x-eov-operation-handler: api
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewUser'
      responses:
        '201':
          description: User created
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      summary: Get users' list
      operationId: getUsers
      x-eov-operation-handler: api
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          description: Start of users name
          required: true
          schema:
            type: string
            minLength: 3
      responses:
        '200':
          description: A list of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsersList"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /identity/{at}:
    get:
      summary: Login
      operationId: login
      x-eov-operation-handler: api
      parameters:
        - name: at
          in: path
          description: Start of users name
          required: true
          schema:
            type: string
            minLength: 3
            maxLength: 125
      responses:
        '200':
          description: Encrypted Bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Identity"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /user/{at}:
    get:
      summary: Get one user informations
      operationId: getOneUser
      x-eov-operation-handler: api
      security:
        - bearerAuth: []
      parameters:
        - name: at
          in: path
          description: Username
          required: true
          schema:
            type: string
            minLength: 3
            maxLength: 125
      responses:
        '200':
          description: One user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: Delete every reference to an account
      description: Incinerate an account entirely
      operationId: incinerate
      x-eov-operation-handler: api
      security:
        - bearerAuth: []
      parameters:
        - in: header
          name: X-msm-Sig
          required: true
          description: SK signature of the body
          schema:
            type: string
            format: byte
            minLength: 172
            maxLength: 172
        - name: at
          in: path
          description: Username
          required: true
          schema:
            type: string
            minLength: 3
            maxLength: 125

      responses:
        '200':
          description: The account doesn't exists anymore
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /contacts:
    put:
      summary: Save the contact list
      operationId: setUserContactList
      x-eov-operation-handler: api
      security:
        - bearerAuth: []
      parameters:
        - in: header
          name: X-msm-Sig
          required: true
          description: SK signature of the body
          schema:
            type: string
            format: byte
            minLength: 172
            maxLength: 172
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Challenge'
      responses:
        '200':
          description: List saved
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /vault:
    put:
      summary: Save the user encrypted secret key
      operationId: setUserVaultItem
      x-eov-operation-handler: api
      security:
        - bearerAuth: []
      parameters:
        - in: header
          name: X-msm-Sig
          required: true
          description: SK signature of the body
          schema:
            type: string
            format: byte
            minLength: 172
            maxLength: 172
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VaultItem'
      responses:
        '200':
          description: Key saved
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: Remove the user encrypted secret key
      operationId: removeUserVaultItem
      x-eov-operation-handler: api
      security:
        - bearerAuth: []
      parameters:
        - in: header
          name: X-msm-Sig
          required: true
          description: SK signature of the body
          schema:
            type: string
            format: byte
            minLength: 172
            maxLength: 172
      responses:
        '200':
          description: Vault item is removed
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /inbox:
    get:
      summary: Get a user inbox
      description: |
        A list of messages headers, the challenge should contain:
        - from
        - sentAt
        - title
      operationId: getInbox
      x-eov-operation-handler: api
      security:
        - bearerAuth: []
      responses:
        '200':
          description: A list of messages headers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inbox"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /message:
    post:
      summary: Send a message
      operationId: writeMessage
      x-eov-operation-handler: api
      security:
        - bearerAuth: []
      parameters:
        - in: header
          name: X-msm-Sig
          required: true
          description: SK signature of the body
          schema:
            type: string
            format: byte
            minLength: 172
            maxLength: 172
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Post'
      responses:
        '201':
          description: Message sent
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /message/{msgId}:
    get:
      summary: Get a message content and delete it from DB
      description: |
        One full message, the challenge should contain:
        - from
        - sentAt
        - title
        - content
      operationId: getMessage
      x-eov-operation-handler: api
      security:
        - bearerAuth: []
      parameters:
        - name: msgId
          in: path
          required: true
          description: The ID of the message to fetch
          schema:
            type: integer
      responses:
        '200':
          description: A complete message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: Force removal of a message
      description: Delete a message from db
      operationId: removeMessage
      x-eov-operation-handler: api
      security:
        - bearerAuth: []
      parameters:
        - in: header
          name: X-msm-Sig
          required: true
          description: SK signature of the body
          schema:
            type: string
            format: byte
            minLength: 172
            maxLength: 172
        - name: msgId
          in: path
          required: true
          description: The ID of the message to delete
          schema:
            type: integer
      responses:
        '200':
          description: The message is removed
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    NewUser:
      type: object
      required:
        - at
        - key
        - signature
        - hash
      properties:
        at:
          type: string
          minLength: 3
          maxLength: 125
        key:
          type: string
          minLength: 788
          maxLength: 788
        signature:
          type: string
          minLength: 268
          maxLength: 268
        hash:
          type: string
          format: byte
          minLength: 172
          maxLength: 172
    Challenge:
      type: object
      required:
        - token
        - passphrase
        - iv
      properties:
        token:
          type: string
          format: byte
        passphrase:
          type: string
          format: byte
          minLength: 684
          maxLength: 684
        iv:
          type: string
          format: byte
          minLength: 24
          maxLength: 24
    Identity:
      allOf:
        - $ref: '#/components/schemas/Challenge'
        - type: object
          properties:
            vault:
              $ref: "#/components/schemas/VaultItem"
    User:
      type: object
      required:
        - id
        - at
        - key
        - signature
      properties:
        id:
          type: string
        at:
          type: string
          minLength: 3
          maxLength: 125
        key:
          type: string
          minLength: 788
          maxLength: 788
        signature:
          type: string
          minLength: 268
          maxLength: 268
    UsersList:
      type: array
      items:
        type: string
        minLength: 3
        maxLength: 125
        
    Inbox:
      type: array
      items:
        $ref: "#/components/schemas/InboxHeader"
    InboxHeader:
      type: object
      required:
        - id
        - challenge
      properties:
        id:
          type: string
        challenge:
          $ref: "#/components/schemas/Challenge"
    Message:
      type: object
      required:
        - id
        - challenge
      properties:
        id:
          type: string
        challenge:
          $ref: "#/components/schemas/Challenge"
    Post:
      type: object
      required:
        - to
        - title
        - content
      properties:
        to:
          type: string
          minLength: 3
          maxLength: 125
        title:
          type: string
          format: byte
          minLength: 684
          maxLength: 684
        content:
          type: string
          format: byte
          minLength: 684
          maxLength: 684
    VaultItem:
      type: object
      required:
        - token
        - iv
      properties:
        token:
          type: string
          format: byte
          minLength: 5548
          maxLength: 5552
        iv:
          type: string
          format: byte
          minLength: 24
          maxLength: 24
    Error:
      type: object
      properties:
        status:
          type: integer
          format: int32
        failedValidation:
          type: boolean
        code:
          type: string
        message:
          type: string